backend/requirements.txt
psycopg2-binary
dotenv
fastapi
uvicorn[standard]
pydantic
email-validator
pyjwt